Update chirp_server default max concurrency.
Currently the chirp_server accepts an unlimited number of connections.  As a
default, this easily allows runaway behavior which can cause a system to
degrade to an unusable state. This recently occurred for the HEP group.  This
commits sets a more reasonable default of 100. As always, users can raise the
concurrency manually using the -M option.

Fixes #468.