Never hang on wait4.
Parrot would block waiting for the parent of a particular thread to wake
up. Such thread, however, was stopped since it was recently created;
additionally, for some reason remained stopped after a SIGCONT. In this
fix, we let parrot continue, and wait on a usleep. This allow to resend
the lost SIGCONT in a further loop iteration.