Hack to wake sleepy threads.
When a thread is newly created using ptrace, parrot is notified of a
SIGSTOP on the thread. Under normal circumstances parrot sends a SIGCONT
to the thread, and we are all happy.

The problem here is that the thread does not receive that SIGCONT. Here
we add newly created threads to a set, and send them SIGCONTs at most
every second until they wake up.