Detect one more flavor of PTRACE_EVENT_STOP.
Apparently there are circumstances where signum == 0 for a PTRACE_EVENT_STOP.
This was wrongly detected as a group-stop which caused the process to hang
because Parrot would not restart it. I believe this only happens when cloning a
thread and we used PTRACE_SEIZE. It didn't show up with our earlier tests
because none of the machines we currently test on are Linux 3.4+.

Maybe related to #502.