Don't allow kernel to raise execve failure.
This is a partial revert of 7a43226fb1628f02650d6d6209cbf488c10cae22.

A bug was found by Haiyan when running the Parrot package run utility where
execve would wrongly succeed. The reason was Parrot would not be able to find a
"/bin/which" executable under its mounted directory (--mount=/=$pkg) but the
kernel *does* find "/bin/which" in the system's root.

This commit goes back to using divert_to_dummy but adds a check to decode_execve
for a "dummy return" so it doesn't try to mess with the result.