msync when changing the channel
As we do in pfs_table.cc when loading a file into the channel, we should msync
when we divert_to_channel.

I also changed M_SYNC to M_ASYNC as it is not necessary that the backing store
be updated synchronously. We only care about the buffer cache being consistent.