Give up control of the terminal.
Parrot used to do all terminal I/O for tracees before the parrot-native-fd
branch was merged. Since then, we allow tracees to natively talk to their
terminal directly with the kernel. As such, we no longer need to hack changing
the foreground process group using tcsetpgrp, when either Parrot or a tracee
receives a terminal stop signal. The root tracee now sets itself as the
foreground process group immediately when it starts, if it has a controlling
terminal.

We also now ignore the terminal stop signals in Parrot. Relevant comment from
the code:

    /* For terminal stop signals, we ignore all.
     *
     * Parrot sometimes writes to the terminal using the debug library, we want
     * to *never* be the foreground process group so we do not call tcsetpgrp
     * to gain control. Instead we just ignore the terminal stop signals. If
     * the terminal is configured with "-tostop" (see stty(1)), then the writes
     * succeed even if we are not the foreground process group. Otherwise, the
     * writes will return with EIO which is fine also. Reading from a terminal
     * requires being the foreground process group (so far which we have never
     * needed to do), we will receive EIO for all reads since we ignore
     * SIGTTIN.
     *
     * SIGTSTOP occurs because the user suspends the foreground process group
     * using ^Z. If this occurs, we also ignore it. We should probably never
     * receive this signal since Parrot becomes a background process group as
     * soon as its root tracee starts.
     */

    install_handler(SIGTSTP,SIG_IGN);
    install_handler(SIGTTIN,SIG_IGN);
    install_handler(SIGTTOU,SIG_IGN);

Fixes #553.