Fix a bug where argv[0] was not set correctly.
Haiyan encountered a problem in a CMS application caused by linking failures
because the application was ultimately using argv[0] to look for libraries
(Python modules in this case). Unfortunately, there's is a bug where we assumed
that ld-linux.so does not replace argv[0]. So what we were doing was:

    execve('ld-linux.so', ['logical-exe', 'physical-exe', ...], [...]);

However, the argv that is ultimately given to the application is:

    ['physical-exe', ...]

what we wanted was:

    ['logical-exe', ...]

When the 'physical-exe' refers to an executable cached by Parrot, this will not
make sense to the application (Python), causing a failure.

Haiyan did most of the work in isolating the cause of this problem.