Allow the "target" fd to be special for fcntl dup.
Jakob Blomer reported a failure by /bin/rm to remove a directory containing a
file. The issue was caused by a failed F_DUPFD on the directory fd. Because of
this, /bin/rm does not iterate over the directory and instead tries to directly
execute rmdir.

The reason for the failed F_DUPFD was an unnecessary check to see if the target
fd was a "special file descriptor" (e.g. the Parrot channel). The target fd would
be the third argument of:

fcntl(fd, F_DUPFD, start);

The "start" fd can be any valid fd as the kernel does not close start fd, it
just uses it as a starting point in the table. [Contrast this with dup2, where
the target fd *is closed* and we *do* need the check.]

Fixes #688.