Add valgrind support.
In order to support valgrind, we have to turn it off prior to tracing the root
tracee. To do this, we use exec. Shipping around another executable to exec is
a bit ridiculous so I'm using the shell instead. The shell requires some signal
handling logic so I'm not enabling this behavior by default. I don't trust it
works everywhere.

Fixes #512.