Atomically open cached file and use open flags.
This corrects a bug where executable files would be opened O_RDWR causing a
ETXTBSY error in a parallel execve in a separate Parrot instance. (This
bug cannot be detected in a single instance of Parrot because Parrot
serializes system calls for its entire tree of tracees.)

Brian Bockelman, Ben Tovar, and Douglas Thain helped locate the cause.

Fixes #742.
Fixes #432.