Edit the executable to change loader.
This commit addreses two different bugs:

o Executing the loader as a wrapper unconditionally changes argv[0]. So prior
  to this commit, Parrot would change the runtime loader by executing:

    execve("/path/to/loader.so", [argv[0], "/path/to/exe", argv[1], argv[2], ...], [...]);

  Unfortunately, the loader will replace argv[0] with "/path/to/exe".
  Applications sensitive to that change will fail. This is issue #678.

o If the loader referred to by the executable (PT_INTERP) is in a non-local
  file system, like cvmfs, then the kernel will not be able execute the program.
  execve will fail with ENOENT. This is issue #781.

This commit solves these problems by editing the PT_INTERP program header of
the executable to refer to a local copy of the loader.

Fixes #678.
Fixes #781.