Use real python/sh.
Apparently on the ND RHEL5 machines, the Python in $PATH refers to a shell script:

    $ which python
    /opt/und/schrodinger/2007/utilities/python

    $ head $(which python)
    #!/bin/sh
    #
    # Find an installation of mmshare in the $SCHRODINGER directory,
    # and launch the application python from the platform-specific
    # bin subdirectory of that installation.   Command-line args are
    # all passed on to python.
    #
    #set -x   # command tracing
    HOST=`uname -n`
    UNAME=`uname`

That's utterly fantastic. Today I learned you can have recursive interpreter
lookups! [That is, for this test, a.py refers to $(which python). So a.py is
interpreted by /opt/und/schrodinger/2007/utilities/python which is interpreted
by /bin/sh.] Today was not a good day.