`umbrella --version` reports the cctools version
m4 treats backtick (`) and apostrophe (') as quote characters, and
removes the quote characters during macro expansion.
Putting the cctools version macro inside umbrella would ruin the content
of umbrella completely due to the usage of backtick and apostrophe.
So the cctools version macro is put in a separate file named cctools_version.