Remove Chirp job file SYMLINK binding.
This binding was added to allow for job sandboxes to be on another file system
while maintaining cheap startup/cleanup (as COPY is naturally expensive). With
commit 8c6065e moving job sandboxes to the Chirp root, this is unnecessary to
support.

It's also worth adding that SYMLINK violated the conceptual design of jobs by
allowing output files to be written to during job execution, rather than at job
completion. It's also possible for applications to readlink the symlink to
obtain the real path of the file under the Chirp root. In principle, we don't
want to allow this.