Adds monitoring to wq master
When using work_queue_enable_monitoring, the first summary of the
resources report corresponds to the master. Resources are polled every 5
minutes. The summary also includes the master command line and the
master project name (if given).

This follows the new convention on the resources archive, in which the
first summary of a file is taken as the parent of all the summaries in
that file.