Do not resolve link target for ACL check.
The link RPC, as with the normal UNIX link syscall, allows for symbolic links
to be hard linked. For example:

    $ ln -s foo bar
    $ ln bar baz
    $ ls -la bar baz
    lrwxrwxrwx 2 batrick users 3 Aug 11 10:45 bar -> foo
    lrwxrwxrwx 2 batrick users 3 Aug 11 10:45 baz -> foo

Currently the Chirp server would perform ACL checks on the resolved target, in
this case foo instead of bar. If this symbolic link points somewhere the
current subject lacks rights, then the RPC will fail with permission denied.

The fix is to use chirp_acl_check_link instead of chirp_acl_check on the
target.

Fixes #887.