Handle log rotations for multiprocess debugging.
Before we used fstat to test the size of the log and then perform a rotation if
it exceeded the maximum. However, this does not handle the situation where the
log was rotated by another process. This would cause the log to be rotated
multiple times with each process working with its own log, all but one
anonymous files.

This commit performs a (racey) stat on the file path for rotation checks and
for testing whether the inode of the file has changed. If the inode has
changed, we reopen the log as it was rotated by a peer process.