Remove old All-Pairs TPDS tree.
Now we have a README which points to an appropriate commit,
82b7a15755068c0047a423ae2b6c2ee88480dd08 (version 2.5.4).

The rationale for this is to reduce the maintenance burden of ensuring the tpds
source is buildable/works with the current version of CCTools. It's also now
possible to point users at a commit since the repository history is now public,
i.e. not in a private subversion repository.

Fixes #961.