Split/delete mappings by whole pages.
munmap operates on whole pages (even if the length is not a multiple of the
page size). Parrot must remove all pages referenced by the [addr, addr+len)
range. This commit computes the affected pages and properly unmaps whole pages.

This commit switches to uintptr_t pointers for correct sizing and arithmetic.