Only munmap if kernel succeeds.
Our code was being more lenient with inputs and would perform an unmap when the
kernel would fail. In particular, our code allowed the address to not be a
multiple of the page size.

The solution is to let the kernel try first and if it succeeds, then we do the
unmap.