Remove wait barrier for some syscalls.
Normally a wait_barrier would be required for accept/pipe/etc. as a new fd may
be added to the fd table. However, because accept may block, the barrier will
potentially cause deadlock. In my tests (TR_parrot_thread_fd.sh), the barrier
is not required as apparently the file descriptors are not added to the file
descriptor table until (after?) the syscall exit event is received.