Test if Parrot is running using parrot_version.
Environment variables are not the right mechanism for testing if Parrot is
already running. In #1203, the PARROT_ENABLED environment variable would be
sent along with jobs to remote systems which would prevent Parrot from running.
Instead, just try a Parrot system call to see if we're already running Parrot.

Fixes #1203.