Do readlink on final exe path component.
This is necessary for the Linux loader which relies on /proc/self/exe being the
canonical path to the executable.

Note that this commit doesn't stricly make /proc/self/exe the canonical path.
We only resolve the final component as that should catch all instances of this
bug. [If there is a link in the middle of the path, it will be resolved
normally in any case.]

Fixes #930.