Reclaim space used by replicas of deleted files.
This commit makes it so the namespace tracks files in use by linking all files
into a special "file store" directory. By examining the link count on each
file, Confuga is able to determine which files have been completely removed
from the namespace. In this way, it can safely and reliably determine which
replicas are in use.

The DeadReplica table is used to create an intent to delete a replica. This is
necessary because a job may be in the process of creating a new replica with
the same RepID (because the replicas are deduplicated by the hash of the
content). For this reason, Confuga must delay unlinking the replica on the
storage node until that storage node is not executing a job.

Fixes #827.