Improve Confuga SN management/membership.
This commit adds a new CLI tool (with accompanying C API) for adding and removing SN in Confuga.

    $ confuga_adm confuga://... sn-add [-p <password file>] [-r <root>] <"uuid"|"address"> <uuid|address>

    and

    $ confuga_adm confuga://... sn-rm <"uuid"|"address"> <uuid|address>

The root directory of the storage node set via the "-r" switch is the location
the HN will use to store SN state. The default root directory is "/.confuga".

The password file passed via "-p" is expected to hold the same contents as a
file on the storage node located at '/password' under the storage node's root
directory set via the '-r' switch. Usually the user doesn't want to explicitly
set this up so the HN will automatically create this password file if the '-p'
option is not used.

Examples:

    $ confuga_adm confuga:///path/to/confuga/ sn-add uuid 46BE3610-42F5-4B7C-B9D1-2B3421C4F172
    OR
    $ confuga_adm confuga:///path/to/confuga/ sn-add address localhost:1024

    $ confuga_adm confuga:///path/to/confuga/ sn-rm uuid 46BE3610-42F5-4B7C-B9D1-2B3421C4F172
    OR
    $ confuga_adm confuga:///path/to/confuga/ sn-rm address localhost:1024

The HN will authenticate a SN by using the SHA1 hash RPC to check the password
file at "/<root>/password" is correct. The HN does not reauthenticate every new
session as sessions are handled at the Chirp reli layer. Instead, the HN
reauthenticates a storage node every 15 minutes.

Storage nodes are now remembered by the HN using UUIDs broadcast as part of
normal Catalog updates. This allows the HN to distinguish between storage nodes
operating with the same address and port (localhost/LAN).

The HN is now also capable of removing storage nodes. This is done by marking a
storage node to be removed which prevents it from being used for future jobs
and new replicas. The HN will remove its replicas so long as sufficient
replicas exist in the cluster. The HN will replicate as necessary so that the
SN can eventually be removed. Once all replicas have been removed and jobs have
completed, the SN will be removed.

Fixes #1324.
Fixes #1325.
Fixes #1326.