configure: replace expr substr with native bash
According to POSIX, `substr` in `expr` produces undefined results. In
some versions of `expr`, it has no special meaning. This causes
configure to fail on these expressions and assign the literal string
"$Format:%.*$" to config commit values, which causes the build to fail
because the closing quote is interpreted as a variable.

Change instances of `expr substr $VAR $index $len` to ${VAR:$index:$len},
which should be supported by all instances of bash.