Don't put resolved symlink in argv[0]
It looks like some previous fixes lined up to cause an issue for the specific
case of exec()ing a symlink to an interpreted script. decode_execve was
overwriting the passed-in path with the value of the symlink. This should
leave argv[0] as passed in while still setting /proc/self/exe to the
canonical path to the executable.