Fix Bug in Symlink Handling #1694
Be more forgiving of symlinks in WQ, if the user provides a directory
containing a broken symlink, or hard links fail in the sandbox.
Fixes the common problem of a transfer failing b/c it contains
a stale editor lock represented as a symlink.

Remove SUNOS ifdefs.

Convert defines to enums.

DEFAULT_WORK_DIR -> DOCKER_WORK_DIR

Fix comment: milliseconds to microseconds.