Rewrite jx_iterate_* functions
I probably wrote that code, but was ugly. These rewritten versions
are cleaner. The reason for rewriting was an issue I ran into with
duplicate keys in a JX object. This is allowed by the JSON spec,
not checked by our implementation, and useful when expressing
argv flags as a JSON object. The previous code could iterate over
keys, but didn't give a way to get the corresponding value.
I initially tried a lookup, but that always returned the *first*
instance of the key, creating a loop in my case. This rewrite adds
functions to get the key/value from the iterator variable.