disconnect worker when it runs out of disk
Before we would try to keep the worker around when a task filled the
disk alloted to the worker. However, this only works for inputs that are
not cached. CJ found that the master would keep sending tasks to a
worker, and those tasks eventually would fill the worker's disk. The
tasks were terminated, but the cached was not cleared, thus disk usage
at the worker kept increasing, and the tasks kept failing.

Since there is no easy way to tell at the worker which files are cached,
this commit makes it so that the worker disconnects.