Change command line interface for ext images again
Rather than forcing the user to deal with service names and labels,
the --ext option now assigns an ID to the service for each ext image
and sets up a mountpoint through Parrot's name resolution system.
The syntax for mounting an ext image is

    --ext IMAGE=MOUNTPOINT

so for example you might say --ext img.ext2=/mnt for an image in the
current directory.