Zero out high 8 bits in ext2fs dirent name_len
The headers give the size field as 16 bits. Since filenames cannot
be longer than 255 bytes, they decided to split the field for
ext2_dir_entry_2 into an 8 bit size field and an 8 bit type.
Based on function signatures, I thought I was dealing with a v1
struct and didn't zero out the high bits of the size;
weirdness ensued.