This is a library on top of work_queue which allows for asynchronous wait.
It should work as drop-in replacement for the classic work_queue
module.

HIGHLY EXPERIMENTAL CODE, IT MAY NOT WORK AT ALL

BEHAVIOUR DIFFERENCE: queue does work even when wait() is not called.

INTERFACE DIFFERENCE: wait(0) returns immediately if no task to return is

It mainly captures submit and wait calls. Tasks submitted get added to a
queue for later real submission. Returned tasks are put into a queue to
be retrieved at the client's convenience. Any other call is made with a
lock (that is, no two work queue calls can be made at the same time.)