Set RPATH for external linkage (HACK)
This is a temporary fix for a linker issue on OSX. Specifically,
Conda-provided libraries are compiled with their install names set
to use @rpath. On OSX, the linked-to library gets to choose the path
that downstream executables use to find them, so all our executables
end up with the Conda-provided paths in their headers. Since we don't
have our RPATH set, this obviously doesn't work.

As a workaround, this commit modifies our library search routines to
set both -L and -rpath for libraries found during configure. For the
specific case of Conda libraries, this works. This is not, however,
a general-purpose solution, and the same issue would crop up again
if someone uses the RPATH slightly differently. This commit
therefore probably shouldn't make it into master, and we need to
figure out a proper solution to this.