Remove JX_FUNCTION type
This was sort of an over-engineered feature, as we had support for
arbitrary lambda expressions in the eval code (but no way to
express/parse that construct). This commit removes the unused
JX function infrastructure, instead opting for a simpler check for
statically-defined functions at call time. Note that this makes it
impossible to bind functions by name (e.g. can't define an alias
for a builtin).

As a side effect, it's no longer necessary to modify the context
passed to `jx_eval()`, so it's safe to skip that copy.