Switch spec format for python_package_analyze
Rather than rolling our own spec format (and needing to add support for
versions, pip + conda packages), just use the format produced by

    conda env export --json

We just filter the dependencies to the ones picked up by the
import search. The keys are sorted, so these specs should be hashable to
identify environments. Note that while the output format is JSON, for
some reason when creating the evironment Conda requires the spec file to
end in .yaml, so choose names (or rename) accordingly.