Support new HTCondor job log format containing year
As of HTCondor version 8.9.3, the default userlog format puts the year
into the job log, looking like this

005 (312.000.000) 2020-03-28 23:01:04

the old format looked like

005 (312.000.000) 03/28 23:01:04

This patch allows cctools to parse both the old and the new format. Note that
the code used to assume the year was 2008, this been updated to 2020, in the
case where we don't know.

Note that you can configure HTCondor to emit the old format by setting
DEFAULT_USERLOG_FORMAT_OPTIONS = legacy