Update python_package_run for persistent environemnts
python_package_run --unpack-to my-env --environment my-env.tar.gz command

will unpack to the directory my-env only once. Further calls to commands
with --unpack-to my-env will therefore run faster.

Calls to python_package_run with the same --unpack-to may be even
simultaneous. The first instance to get a writing lock untars the file.
Other instances will use the environment found in the directory
'my-env'. This is quite advantegeous in the case of a work_queue_worker
running several tasks at once, as not all tasks will hit the filesystem
simultaneously unpacking the tarball.

If --unpack-to is not given, then the previous behaviour of creating a
temporary directory that is removed at the end of execution is followed.

Note that now the environment is specified with --environment. This is a
required argument.