Persistent environments with python_package_run #2282
Update python_package_run for persistent environemnts

python_package_run --unpack-to my-env --environment my-env.tar.gz command

will unpack to the directory my-env only once. Further calls to commands
with --unpack-to my-env will therefore run faster.

Calls to python_package_run with the same --unpack-to may be even
simultaneous. The first instance to get a writing lock untars the file.
Other instances will use the environment found in the directory
'my-env'. This is quite advantegeous in the case of a work_queue_worker
running several tasks at once, as not all tasks will hit the filesystem
simultaneously unpacking the tarball.

If --unpack-to is not given, then the previous behaviour of creating a
temporary directory that is removed at the end of execution is followed.

Note that now the environment is specified with --environment. This is a
required argument.

added licenses

Move to python_packaging to python_environments

We already have a `packaging` directory, which is related to the
distribution of cctools.

Install packaging scripts to cctools/bin

This is needed for @zsurma's coffea work queue executor. Rather than
including python_package_run as part of the executor, with this commit
it will become part of the cctools installed by conda.