Increase the minimum file transfer time to 60 seconds. (#2295)
Rationale: Although 10 seconds ought to be plenty for an epsilon-sized file,
consider what happens when streaming a large directory.  A large set of files
gets pushed over to the worker, which stuffs everything in the buffer/cache.
Suddenly, when the buffer cache is full, transfers stutter as data is pushed
to disk.  The transfer time of a single item may be affected as much by what
came before, than by its own size.

An even better solution would take into account the time to transfer everything
prior to this item.  But until we have that more sophisticated approach, this
change will make things a little more stable.