rmonitor: Remove race condition when detecting root process #2308
For long-running subprocesses (i.e., running for more than .25s), the
resource_monitor measures the resources one last time before the process
exits. This is not done for short-running subprocesses, as the measurement
overhead becomes noticeable (e.g., madgraph in HEP, that internally uses
make, gcc, and a series of very small python and shell scripts).

However, this distinction is only relevant for subprocesses, and not for
the very first process created by the application to be measured. A
measurement before exit is always done for this first process, as this
guarantees a measurement of any of the short-running subprocesses via
getrusage.

The previous code that labeled a process as the first process had a race
condition in which several processes may identify themselves as the root
process, with the potential of a noticeable overhead. This commit fixes
this by labeling the first process before the exec call, rather than
using the ld preloaded monitoring helper library to label the correct
process.

The incorrect behaviour was observed by running scripts
with  #!/bin/bash. (This was not observed with !#/bin/sh.)