Fixes third-put/get, updates to %% replacement.
Adds work_queue_task_file_command to work_queue.h. (Function existed in .c file.)

Changed cmd so that %% is replaced with cache name at worker. Before the
contents of the file were read/writen from stdin/stdout, which was not
supported by all commands.

Implementation bug: I did not change the original implementation when
dealing with errors. If transfer command fails, worker disconnects and
transfer is tried in another worker. If command is incorrect, or the
file is not accessible, this will disconnect all workers. Ideally, we
would simply fail or retry the task somewhere else.