Shrink MQ header
With a per-frame size limit (compile-time defined to 1<<16 bytes),
there's no need to send a full 64 bits in the header. Switching
that to a 32 bit value (could recompile with gigabyte-sized frames,
if someone really wants to do that), shortening the magic bytes,
and reducing the padding lets us squeeze the header into 8 bytes.
That gives 0.01% overhead relative to just sending the bytes
over the wire, which can be reduced as far as 3.7e-9 by compiling
with MQ_FRAME_WIDTH increased.