wq factory: allow abort while sleep
This commit checks abort_flag each second the factory sleeps. This
allows a much faster response when using the factory with PythonTask.

The previous behaviour made the factory sleep for a whole period before
checking the abort_flag. Fast executing PythonTask would then have a 30s
overhead when using the 'with workers:' syntax.