wq worker: cleanup workspace after complete disconnect from manager
Otherwise, there was a race condition where a task could produce a file
between the removal of the workspace and the termination of the task.
This issue only happened for tasks that wrote files in the workspace
outside their sandbox. (Which breaks the wq task abstraction and
shouldn't really happen but for very special conditions.)