rmonitor: fix --pid mode
With the changes in 360a6da24a7, the monitor was forwarding terminations
signals to the main monitored process, even when attaching to an already
running process. This commit makes it so that termination signals
are not forwarded to the main process in such case.

Further, to handle unresponsive processes in the general case, the
monitor now terminates itself and any known process if two termination
signals are received in less than a second.