vine, wq: fix double free where tasks were freed before the manager
Quirk of python where finalizers are called in the order they were
defined on exit, regardless of the reference count. If there were active
tasks, this meant that the finalizer of a task would free it, and then
the manager would try to do the same from C.

Now tasks check if the manager finalizer is still alive, and whether the
manager knows about the task.