adds ramp-down mode
When enabled (aq.tune("ramp-down-heuristic", 1)), and the number of waiting tasks is less than the number of
connected workers, then tasks are allocated all the free resources of a
worker. This avoids tasks terminated with resource exhaustion when there
are more resources than tasks.

If monitoring is not enabled, or if it just set to measure, then this
heuristic has no effect. By default it is disabled.