dask+vine: Do not immediately serialize waiting calls.
Since serialization may be expensive, with this commit we set a maximum
number of tasks to submit to the manager per wait call. The rest of the
tasks are added to a FIFO queue to wait their turn to be serialized.
This has the effect of overlapping serialization with execution.