vine: use a single file for wrappers, buffer for functions
Previously TaskVine would generate wrapper and function file per task,
even when the wrapper was the same for all the PythonTask, and the
function was used by other tasks.

This pr makes it so that wrappers and functions are serialized once.
Further, functions are not written to disc but kept as a declare_buffer.