vine: fix creation error message (#4131)
When taskvine could not create a manager because the logs or staging path
are incorrect, we would print a message that the manager fail to attach
to port 0. This was not informative.

This change makes it so that we print the log and staging path, and
print the original value of the port range explicitely given.